Made it work...

+++ HX711-multi.h
@@ -3,6 +3,7 @@
 
 #if ARDUINO >= 100
 #include "Arduino.h"
+#include <Streaming.h>
 #else
 #include "WProgram.h"
 #endif
@@ -25,7 +26,7 @@ class HX711MULTI
 		// channel selection is made by passing the appropriate gain: 128 or 64 for channel A, 32 for channel B
 		// count: the number of channels
 		// dout: an array of pin numbers, of length 'count', one entry per channel
-		HX711MULTI(int count, byte *dout, byte pd_sck, byte gain = 128);
+		HX711MULTI(byte count, byte *dout, byte pd_sck, byte gain = 128);
 
 		virtual ~HX711MULTI();
 
@@ -43,10 +44,10 @@ class HX711MULTI
 		void set_gain(byte gain = 128);
 
 		// waits for the chip to be ready and returns a reading
-		void read(long *result = NULL);
+		void read(long* result = NULL);
 
 		// same as read, but does not offset the values according to the tare
-		void readRaw(long *result = NULL);
+		void readRaw(long* result = NULL);
 
 		// set the OFFSET value for tare weight
 		// times: how many times to read the tare value
